# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API de chat com autenticação de usuários.",
    "version": "1.0.0",
    "title": "Chat API"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "auth",
      "description": "Operações sobre autenticação de usuários"
    },
    {
      "name": "chat",
      "description": "Operações sobre chats"
    },
    {
      "name": "message",
      "description": "Operações sobre mensagens"
    },
    {
      "name": "article",
      "description": "Operações sobre artigos"
    },
    {
      "name": "category",
      "description": "Operações sobre categorias"
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/authenticate": {
      "post": {
        "tags": ["auth"],
        "summary": "Autenticar usuário e retornar JWT",
        "description": "Retorna um token JWT para usuários autenticados",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Credenciais de login do usuário",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["emailOrUsername", "password"],
              "properties": {
                "emailOrUsername": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticação bem-sucedida",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "Authorization": {
                "type": "string",
                "description": "Bearer {token}"
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          },
          "400": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": ["auth"],
        "summary": "Criar uma nova conta de usuário",
        "description": "Cria uma nova conta de usuário.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados de criação do usuário",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "username", "email", "type", "password"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          }
        }
      }
    },
    "/chat/create": {
      "post": {
        "tags": ["chat"],
        "summary": "Criar um novo chat",
        "description": "Cria um novo chat entre um atendente e um cliente.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados do chat",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "attendantId", "clientId"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "attendantId": {
                  "type": "integer"
                },
                "clientId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Chat criado com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/chats": {
      "get": {
        "tags": ["chat"],
        "summary": "Listar todos os chats",
        "description": "Lista todos os chats.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de chats",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "clientId": {
                    "type": "integer"
                  },
                  "attendantId": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/messages": {
      "post": {
        "tags": ["message"],
        "summary": "Enviar uma nova mensagem",
        "description": "Envia uma nova mensagem em um chat específico.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados da mensagem",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["chatId", "content"],
              "properties": {
                "userId": {
                  "type": "integer"
                },
                "chatId": {
                  "type": "integer"
                },
                "content": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Mensagem enviada com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/chat/{chatId}/messages": {
      "get": {
        "tags": ["message"],
        "summary": "Listar mensagens de um chat",
        "description": "Lista todas as mensagens de um chat específico.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do chat"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensagens",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "chatId": {
                    "type": "integer"
                  },
                  "userId": {
                    "type": "integer"
                  },
                  "content": {
                    "type": "string"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/articles/save": {
      "post": {
        "tags": ["article"],
        "summary": "Salvar um novo artigo",
        "description": "Cria um novo artigo.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados do artigo",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "title",
                "body",
                "category",
                "summary",
                "coverImage"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "body": {
                  "type": "string"
                },
                "category": {
                  "type": "integer"
                },
                "summary": {
                  "type": "string"
                },
                "coverImage": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Artigo salvo com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/articles": {
      "get": {
        "tags": ["article"],
        "summary": "Listar todos os artigos",
        "description": "Lista todos os artigos.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de artigos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/categories/save": {
      "post": {
        "tags": ["category"],
        "summary": "Salvar uma nova categoria",
        "description": "Cria uma nova categoria.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados da categoria",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["title"],
              "properties": {
                "title": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Categoria salva com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/categories": {
      "get": {
        "tags": ["category"],
        "summary": "Listar todas as categorias",
        "description": "Lista todas as categorias.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira seu token de portador no formato **Bearer <token>**"
    }
  }
}